Fixed a number of things
I opened this branch so long ago, I can't remember what the specific objectives were, but I know what I accomplished. So…

One of the absolute elements in the "banner" section of the page was not positioning relative to the correct ancestor. I fixed that. I modified the navigation bar so that it will stick to the top of the browser window when the page is scrolled.

Considering how much time it took, I'm surprised at how little I accomplished. Typically, however, efforts to fix one problem creates new ones.